refactoring: replaced pieces.t.ts#PieceType with compressed.types.t.ts#PieceType
