+++ angular-chess/src/app/components/chess-board.component.ts
@@ -8,8 +8,8 @@ import { HighlightingService } from '../services/highlighting.service';
 import { MoveHistoryService } from '../services/move-history.service';
 import { PositioningService } from '../services/positioning.service';
 import { Board, HighlightColor, Result, SquareWithHighlight } from '../types/board.t';
-import { Color, Square } from '../types/compressed.types.t';
-import { Move, Piece, PieceType } from '../types/pieces.t';
+import { Color, PieceType, Square } from '../types/compressed.types.t';
+import { Move, Piece } from '../types/pieces.t';
 import BoardUtils from '../utils/board.utils';
 import LoggingUtils, { LogLevel } from '../utils/logging.utils';
 import MoveGenerationUtils from '../utils/move-generation/move.generation.utils';
