+++ angular-chess/src/app/services/chess-board.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
-import { Color, Square } from '../types/compressed.types.t';
-import { Piece, PieceType } from '../types/pieces.t';
+import { Color, PieceType, Square } from '../types/compressed.types.t';
+import { Piece } from '../types/pieces.t';
 import { ChessBoardService } from './chess-board.service';
