+++ angular-chess/src/app/utils/board.utils.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from "@angular/core/testing";
 import { Board, Result } from "../types/board.t";
-import { Color, Direction, Square } from "../types/compressed.types.t";
-import { Move, Piece, PieceType } from "../types/pieces.t";
+import { Color, Direction, PieceType, Square } from "../types/compressed.types.t";
+import { Move, Piece } from "../types/pieces.t";
 import BoardUtils from "./board.utils";
 import TestUtils from "./test.utils";
