+++ angular-chess/src/app/utils/engine.utils.spec.ts
@@ -1,6 +1,6 @@
 import { Board } from "../types/board.t";
-import { Color, Square } from "../types/compressed.types.t";
-import { Move, Piece, PieceType } from "../types/pieces.t";
+import { Color, PieceType, Square } from "../types/compressed.types.t";
+import { Move, Piece } from "../types/pieces.t";
 import BoardUtils from "./board.utils";
 import CopyUtils from "./copy.utils";
 import EngineUtils, { MoveWithScore } from "./engine.utils";
