+++ angular-chess/src/app/utils/move-execution.utils.spec.ts
@@ -1,6 +1,6 @@
 import { Board } from "../types/board.t";
-import { Color, Square } from "../types/compressed.types.t";
-import { Move, PieceType } from "../types/pieces.t";
+import { Color, PieceType, Square } from "../types/compressed.types.t";
+import { Move } from "../types/pieces.t";
 import BoardUtils from "./board.utils";
 import MoveExecutionUtils from "./move-execution.utils";
 import TestUtils from "./test.utils";
