+++ angular-chess/src/app/utils/move-generation/move-generation.knight.handler.ts
@@ -1,6 +1,6 @@
 import { Board, } from "src/app/types/board.t";
-import { Square } from "src/app/types/compressed.types.t";
-import { Move, Piece, PieceType } from "src/app/types/pieces.t";
+import { PieceType, Square } from "src/app/types/compressed.types.t";
+import { Move, Piece } from "src/app/types/pieces.t";
 import BoardUtils from "../board.utils";
 import PieceUtils from "../piece.utils";
 import SquareUtils from "../square.utils";
