+++ angular-chess/src/app/utils/move-generation/move-generation.pawn.handler.spec.ts
@@ -1,6 +1,6 @@
 import { Board } from "src/app/types/board.t";
-import { Color, Square } from "src/app/types/compressed.types.t";
-import { Piece, PieceType } from "src/app/types/pieces.t";
+import { Color, PieceType, Square } from "src/app/types/compressed.types.t";
+import { Piece } from "src/app/types/pieces.t";
 import BoardUtils from "../board.utils";
 import SquareUtils from "../square.utils";
 import { MoveGenerationPawnHandler } from "./move-generation.pawn.handler";
