+++ angular-chess/src/app/utils/move-generation/move-generation.queen.handler.ts
@@ -1,6 +1,6 @@
 import { Board } from "src/app/types/board.t";
-import { Direction, Square } from "src/app/types/compressed.types.t";
-import { Move, Piece, PieceType } from "src/app/types/pieces.t";
+import { Direction, PieceType, Square } from "src/app/types/compressed.types.t";
+import { Move, Piece } from "src/app/types/pieces.t";
 import BoardUtils from "../board.utils";
 import CopyUtils from "../copy.utils";
 import SquareUtils from "../square.utils";
