+++ angular-chess/src/app/utils/move-generation/move.generation.utils.ts
@@ -1,6 +1,6 @@
 import { Board } from "src/app/types/board.t";
-import { Color, Square } from "src/app/types/compressed.types.t";
-import { Move, Piece, PieceType } from "src/app/types/pieces.t";
+import { Color, PieceType, Square } from "src/app/types/compressed.types.t";
+import { Move, Piece } from "src/app/types/pieces.t";
 import LoggingUtils, { LogLevel } from "../logging.utils";
 import SquareUtils from "../square.utils";
 import { MoveGenerationBishopHandler } from "./move-generation.bishop.handler";
