+++ angular-chess/src/app/utils/move.search.utils.spec.ts
@@ -1,6 +1,6 @@
 import { Board } from "../types/board.t";
-import { Color, Square } from "../types/compressed.types.t";
-import { Move, PieceType } from "../types/pieces.t";
+import { Color, PieceType, Square } from "../types/compressed.types.t";
+import { Move } from "../types/pieces.t";
 import BoardUtils from "./board.utils";
 import MoveSearchUtils from "./move.search.utils";
 import TestUtils from "./test.utils";
