+++ angular-chess/src/app/utils/square.utils.spec.ts
@@ -1,6 +1,6 @@
 import { Board } from '../types/board.t';
-import { Color, Direction, File, Rank, Square } from '../types/compressed.types.t';
-import { Move, Piece, PieceType } from '../types/pieces.t';
+import { Color, Direction, File, PieceType, Rank, Square } from '../types/compressed.types.t';
+import { Move, Piece } from '../types/pieces.t';
 import SquareUtils, { Position } from './square.utils';
 
 describe('SquareUtils', () => {
